pyproject.toml
[tool.poetry]
name = "llm-app"
version = "0.1.0"
description = ""
authors = ["Dat Le <minh_dat_le@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-dotenv = "^1.0.1"
bcrypt = "^4.1.3"
pyjwt = "^2.8.0"
langchain = "^0.2.9"
langchain-community = "^0.2.7"
langchain-openai = "^0.1.17"
langchain-google-genai = "^1.0.7"
langchain-anthropic = "^0.1.20"
langchain-huggingface = "^0.0.3"
sentencepiece = "^0.2.0"
langchainhub = "^0.1.20"
text-generation = "^0.7.0"
langchain-ai21 = "^0.1.7"
beautifulsoup4 = "^4.12.3"
pymupdf = "^1.24.7"
easyocr = "^1.7.1"
python-bidi = "^0.5.0"
pytesseract = "^0.3.10"
langchain-chroma = "^0.1.2"
tqdm = "^4.66.4"
pypdf = "^4.3.1"
sentence-transformers = "^3.0.1"
faiss-cpu = "^1.8.0.post1"
slack-bolt = "^1.19.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
